refactor: api 명세 변경 msw 반영 및 singer 핸들러 분리
