+++ frontend/src/mocks/handlers/index.ts
@@ -1,7 +1,8 @@
 import memberHandlers from './memberHandlers';
 import searchHandlers from './searchHandlers';
+import singerHandlers from './singerHandlers';
 import songsHandlers from './songsHandlers';
 
-const handlers = [...memberHandlers, ...searchHandlers, ...songsHandlers];
+const handlers = [...memberHandlers, ...searchHandlers, ...singerHandlers, ...songsHandlers];
 
 export default handlers;
